Added PUT methods to FN125Tags, FN126, and FN127 controllers
this commit adds a put method to the FN125Tag, FN126 and Fn127
controller so that whole objects can be replaced (which is more
consistent with how data will be coming from many clients).  It also
removes the key field arguements in the POST functions. These were not
being used and made the url unecessarily long.