FishnetDataTables and pydantic models for FN125Tag and FN125Lamprey
This commit added FishnetDataTable instances for FN125Tag and
FN125Lamprey data.  This commit also updates all of the pydantic
models to version 2.0 and removes all of the sql files for models that
have been converted to this point. the Ping scripts and pytests all
continue to work as expected.