Updated FN125 to use FishnetTable and Pydandic model
This commit updates the FN125 controller to use a FishnetTable class to
generate the sql required for crud queries. It also uses a pydantic
model to validate the data. Tests are included to ensure that the
pydantic model works as expected.  Small tweaks and updates to other
files. All tests continue to pass.