Added values method to FishnetTable class. Updated all instances
this commit adds a values method to the FishnetTable class that will
ensure values passed to the sql clauses match the order of the fields.
A test case is included. This method supersedes the get_data_values
function which has been removed from the code base. All tests pass and
ping scripts run to completion.